requirements.txt
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mistune==3.0.2
packaging==23.2
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
six==1.16.0
smmap==5.0.1
urllib3==2.0.7
Werkzeug==3.0.1
zipp==3.17.0
